fixed non-XML comment related errors
