+++ Revoke.NET.MongoDB/RevokeService.cs
@@ -8,7 +8,7 @@ public static class RevokeService
     public static IServiceCollection AddRevokeMongoStore(this IServiceCollection services)
     {
         return services.AddSingleton(
-            provider => MongoBlackList.CreateStoreAsync("RevokeStore", MongoClientSettings.FromConnectionString("mongodb://127.0.0.1:27017/RevokeStore"))
+            _ => MongoBlackList.CreateStoreAsync("RevokeStore", MongoClientSettings.FromConnectionString("mongodb://127.0.0.1:27017/RevokeStore"))
                 .GetAwaiter()
                 .GetResult());
     }
@@ -16,7 +16,7 @@ public static class RevokeService
     public static IServiceCollection AddRevokeMongoStore(this IServiceCollection services, string dbName, MongoClientSettings settings)
     {
         return services.AddSingleton(
-            provider => MongoBlackList.CreateStoreAsync(dbName, settings)
+            _ => MongoBlackList.CreateStoreAsync(dbName, settings)
                 .GetAwaiter()
                 .GetResult());
     }
