+++ Revoke.NET.Redis/RevokeService.cs
@@ -7,7 +7,7 @@ public static class RevokeService
     public static IServiceCollection AddRevokeRedisStore(this IServiceCollection services)
     {
         return services.AddSingleton(
-            provider => RedisBlackList.CreateStoreAsync("127.0.0.1:6379")
+            _ => RedisBlackList.CreateStoreAsync("127.0.0.1:6379")
                 .GetAwaiter()
                 .GetResult());
     }
@@ -15,7 +15,7 @@ public static class RevokeService
     public static IServiceCollection AddRevokeRedisStore(this IServiceCollection services, string connectionString)
     {
         return services.AddSingleton(
-            provider => RedisBlackList.CreateStoreAsync(connectionString)
+            _ => RedisBlackList.CreateStoreAsync(connectionString)
                 .GetAwaiter()
                 .GetResult());
     }
